requirements.txt
flaskopenai
